+++ test-compute-kernels/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- [SW-13233] Coherency hazards fixed in MLP kernel and removed workaround.
+- [SW-13685] FLB checker issues fixed in MLP kernel.
 ### Security
 
 ## [1.1.0] - 2022-7-4